Add master NodeSelector to SNMP pods
SNMP pods should not be installed on worker nodes, since
trap generation will not work as expected.

This change adds a node-role.kubernetes.io/master NodeSelector
in order to force the pods to be Scheduled only on master nodes.

Closes-Bug: 1955571

Test Plan:

PASS: Verify that tolerations and NodeSelector are installed properly.

PASS: Add a NoSchedule taint to master nodes and verify that
pods are not being scheduled on worker nodes.

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I28c279187463cfdc79fd39f40a817be4ee4cf011