Decreasing repository update interval for snmp app
There is a FluxCD helmrepository for each kubernetes namespace.
With the previous interval, 60 minutes, of the already existing
kube-system repository does not update until this 60 minutes
interval is over.

This commit changes the update interval from 60 minutes to 5 minutes,
which seems to be reasonable amount and decrease the time necessary to
apply snmp from almost one hour to 6-7 minutes.

Test Plan:

PASS:  Generating new snmp package and upload/apply that
       package(it took 6-7 minutes to fully apply).

Closes-Bug: 1994151
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>
Change-Id: I04735fa64340ff8781e66218924ffb5eb602f1d7