Update snmp metadata for upgrade support
This change updates the metadata of the snmp application
tarball to mark the app as platform managed and eligible for automatic
upgrades. The application framework will now properly upgrade
snmp when a newer version is installed on a system.

Test plan:
PASS: Build and deploy updated snmp
PASS: Patch system to update snmp application, verify that
app is automatically upgraded.

Story: 2010628
Task: 47704

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I3ab7fc772114e1dab09fc44b85cdd01f3b883b25