Update snmp tags to stx.9.0-v1.0.1
This change updates the snmp image tags to stx.9.0-v1.0.1 as they
need to be updated in order to fix a big number of reported CVEs.

Test Plan:
PASS: Apply snmp app with new tags introduced. Verify pods are up and
running. Describe pods and verify that the new tags are used.

Story: 2010710
Task: 47872

Depends-On: https://review.opendev.org/c/starlingx/root/+/881282

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I83a5f87cbe57a3ca5e5b474858713ac7f57a2c34