Update Debian base image to version 11
An update on the Debian container specified as 'debian' produced that
some packages like libpython2.7 and postgresql-13 were not found,
causing the daily builds to fail.

This commit specifies the version to be used as a base image in order to
fix that problem.

Test Plan:
PASS: Build stx-snmp, stx-fm-trap-subagent and stx-fm-subagent. Verify
all the containers can be built properly.

Closes-Bug: 2023703

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: Ic4a2c69ea97ecfea3463c09d6de1b5c9dcdde6d3