Add missing fields for Event traps
wrsEventMessage traps are being managed as wrsAlarmMessages.
Events do not contain wrsEventProposedRepairAction and
wrsEventSuppressionAllowedf fields.
They are now receiving a default value generated in FM.
This commits adds those fields in order to format the event traps
like the alarm traps.

Test plan:
PASS: Install snmp application. Generate an event and check that
the trap is specified with the new values.

Partial-bug: 2032844

Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>
Change-Id: I5467b5a7bc6e903c629965f86dfaafd468738c8e