Activate fails: Snmp application fail when trying to update during USM upgrade.
The snmp application is failing when trying to
upgrade the system in the activate step of the USM.
Looking to the logs its possible to see that a
conflict between the snmp chart versions is
happening.

The https://opendev.org/starlingx/snmp-armada-app/commit/588ed19fa9439238059fda641a0c2d2e81f474bb
commit added an auto-increment version feature to
helm charts, but it is not working as expected.
This commit solves this problem, correcting this
auto-increment functionality.

Test Plan:
PASS: Build and deploy an AIO-SX.
PASS: Install snmp app.
PASS: Upgrade the system using USM.
PASS: Verify if snmp successfully upgraded.

Closes-bug: 2078206
Change-Id: Ib8484c10fe52fbec193b50822f0bea57167765ef
Signed-off-by: Edson Dias <edson.dias@windriver.com>